requirements.txt
numpy
opencv-python
tensorflow-macos
tensorflow-metal
keras
Pillow
scipy
matplotlib
scikit-learn
scikit-image
tqdm
pandas
seaborn
imageio
requests
beautifulsoup4
lxml
PyYAML
h5py
joblib
pydub
soundfile
librosa
pycocotools
pyyaml
pycocotools
icrawler
streamlit